lib: rename textfile_copy to file_copy
Because it does not depend on the file being a text file.
